Move nsxt connection part to a separate dir
