+++ pkg/cloudprovider/vsphere/loadbalancer/config/types_yaml.go
@@ -16,6 +16,10 @@
 
 package config
 
+import (
+	nsxtcfg "k8s.io/cloud-provider-vsphere/pkg/nsxt/config"
+)
+
 /*
 	TODO:
 	When the INI based cloud-config is deprecated, this file should be renamed
@@ -31,7 +35,7 @@ package config
 type LBConfigYAML struct {
 	LoadBalancer      LoadBalancerConfigYAML                  `yaml:"loadBalancer"`
 	LoadBalancerClass map[string]*LoadBalancerClassConfigYAML `yaml:"loadBalancerClass"`
-	NSXT              NsxtConfigYAML                          `yaml:"nsxt"`
+	NSXT              nsxtcfg.NsxtConfigYAML                  `yaml:"nsxt"`
 }
 
 // LoadBalancerConfigYAML contains the configuration for the load balancer itself
@@ -59,24 +63,4 @@ type LoadBalancerClassConfigYAML struct {
 	TCPAppProfilePath string `yaml:"tcpAppProfilePath"`
 	UDPAppProfileName string `yaml:"udpAppProfileName"`
 	UDPAppProfilePath string `yaml:"udpAppProfilePath"`
-}
-
-// NsxtConfigYAML contains the NSX-T specific configuration
-type NsxtConfigYAML struct {
-	// NSX-T username.
-	User string `yaml:"user"`
-	// NSX-T password in clear text.
-	Password string `yaml:"password"`
-	// NSX-T host.
-	Host string `yaml:"host"`
-	// InsecureFlag is to be set to true if NSX-T uses self-signed cert.
-	InsecureFlag bool `yaml:"insecureFlag"`
-	// RemoteAuth is to be set to true if NSX-T uses remote authentication (authentication done through the vIDM).
-	RemoteAuth bool `yaml:"remoteAuth"`
-
-	VMCAccessToken     string `yaml:"vmcAccessToken"`
-	VMCAuthHost        string `yaml:"vmcAuthHost"`
-	ClientAuthCertFile string `yaml:"clientAuthCertFile"`
-	ClientAuthKeyFile  string `yaml:"clientAuthKeyFile"`
-	CAFile             string `yaml:"caFile"`
 }
